# swagger.yaml
openapi: 3.0.3
info:
  title: Absensi Selfie API
  version: 2.0.0
  contact:
    name: Mohamad Radisha
    email: pr0ph0z23@gmail.com
servers:
  - url: https://absensiselfie-api.pptik.id:5026
    description: Pre-production/staging server
  - url: https://pptik-local.pptik.id:5026
    description: Private development server
  - url: https://raspi.pptik.id:5026
    description: Private development server
  - url: http://api.absensi-selfie.pptik.id:5006
    description: Production server, this is used in API version less than 2.0.0
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    SuccessResponse:
      type: object
      properties:
        code:
          type: integer
          format: int32
          example: 200
        status:
          type: boolean
          example: true
        message:
          type: string
          example: Success
...
    UnprocessableEntityResponse:
      type: object
      properties:
        code:
          type: integer
          format: int32
          example: 422
        status:
          type: boolean
          example: false
        message:
          type: string
          example: The request you sent is unable to process
...
    UserListResponse:
      allOf:
        - $ref: "#/components/schemas/SuccessResponse"
        - type: object
          properties:
            data:
              type: object
              properties:
                users:
                  type: array
                  items:
                    $ref: "#/components/schemas/UserList"
                number_of_pages:
                  type: number
...
    UserRegister:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
          format: password
        position:
          type: string
        id_card:
          type: string
        company:
          type: string
          description: Company code
        local_image:
          type: string
        unit:
          type: string
        phone_number:
          type: string
        image:
          type: string
          format: binary
        identity_card_image:
          type: string
          format: binary
        family_card_image:
          type: string
          format: binary
        user_location_address:
          type: string
        user_location_long:
          type: number
        user_location_lat:
          type: number
        user_property_location_long:
          type: number
        user_property_location_lat:
          type: number
        role:
          type: string
...
    UserList:
      type: object
      properties:
        PHONE_NUMBER:
          type: string
        NAME:
          type: string
        EMAIL:
          type: string
        POSITION:
          type: string
        ID_CARD:
          type: string
        COMPANY:
          type: string
        IMAGE:
          type: string
        UNIT:
          type: string
        GUID:
          type: string
        IDENTITY_CARD_IMAGE:
          type: string
        FAMILY_CARD_IMAGE:
          type: string
...
paths:
...
  /admin/user:
    get:
      tags:
        - Admin Users
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
      responses:
        200:
          description: Get user list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserListResponse"
    post:
      tags:
        - Admin Users
      security:
        - bearerAuth: []
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/UserRegister"
      responses:
        200:
          description: Register success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        422:
          description: There are two possible messages. You can produce the `No company with that code` error with wrong company code (in the `company` field) and `E-mail already registered` error with registered email
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnprocessableEntityResponse"
...